\name{GoogleAPI}
\alias{GoogleAPI}
\title{R interface for Google Civic API}
\usage{
GoogleAPI(method, region = "korea", key = getOption("polrGoogleKey"))
}
\description{R interface for Google Civic API}
\arguments{
  \item{method}{method chooses }
  \item{region}{region is for }
  \item{key}{key is the target}
}

\value{
}
\details{
See
\url{(https://developers.google.com/civic-information/)}
}
\seealso{ \code{PopongAPI}, \code{SunlightgAPI}  }

\examples{

}

\keyword{API}
