% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SunlightAPI}
\alias{SunlightAPI}
\title{R interface for Sunlight API}
\usage{
SunlightAPI(method, key = getOption("polrSunlightKey"))
}
\arguments{
\item{method}{method to call}

\item{key}{API key}
}
\description{
See \url{(https://sunlightlabs.github.io/congress/)}
}
\seealso{
\code{\link{GoogleAPI}}, \code{\link{PopongAP}}
}
\keyword{API}

