% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GoogleAPI}
\alias{GoogleAPI}
\title{R interface for Google Civic API}
\usage{
GoogleAPI(method, key = getOption("polrGoogleKey"))
}
\arguments{
\item{method}{method to call}

\item{key}{API key}
}
\description{
See \url{(https://developers.google.com/civic-information/)}
}
\seealso{
\code{\link{PopongAPI}}, \code{\link{SunlightAPI}}
}
\keyword{API}

