\name{GoogleAPI}
\alias{GoogleAPI}
\title{R interface for Google Civic API}
\usage{
GoogleAPI(method, key = getOption("polrGoogleKey"))
}
\arguments{
  \item{method}{method to call}

  \item{key}{API key}
}
\description{
See
\url{(https://developers.google.com/civic-information/)}
}
\examples{
TBD
}
\seealso{
\code{\link{PopongAPI}}, \code{\link{SunlightAPI}}
}
\keyword{API}

