\name{PopongAPI}
\alias{PopongAPI}
\title{R interface for Popong API}
\usage{
PopongAPI(method, region = "korea", key = getOption("polrPopongKey"))
}
\description{R interface for Korea Political Data Retriever (Popong) API}
\arguments{
  \item{method}{method chooses }
  \item{region}{region is for }
  \item{key}{key is the target}
}

\value{
}
\details{
See
\url{(http//popong.com)}
}
\seealso{ \code{GoogleAPI}, \code{SunlightgAPI}  }

\examples{

}

\keyword{API}
