\name{PopongAPI}
\alias{PopongAPI}
\title{R interface for Popong API}
\usage{
PopongAPI(method, key = getOption("polrPopongKey"), debug = FALSE)
}
\arguments{
  \item{method}{method to call}

  \item{key}{API key}

  \item{debug}{Enable debugging mode}
}
\description{
See \url{(http://data.popong.com)}
}
\seealso{
\code{\link{GoogleAPI}}, \code{\link{SunlightAPI}}
}
\keyword{API}

