% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PopongAPI}
\alias{PopongAPI}
\title{R interface for Popong API}
\usage{
PopongAPI(method, key = getOption("polrPopongKey"))
}
\arguments{
\item{method}{method to call}

\item{key}{API key}
}
\description{
See \url{(http://data.popong.com)}
}
\seealso{
\code{\link{GoogleAPI}}, \code{\link{SunlightAPI}}
}
\keyword{API}

