\name{SunlightAPI}
\alias{SunlightAPI}
\title{R interface for Sunlight API}
\usage{
SunlightAPI(method, key = getOption("polrSunlightKey"), debug = FALSE)
}
\arguments{
  \item{method}{method to call}

  \item{key}{API key}

  \item{debug}{Enable debugging mode}
}
\description{
See \url{(https://sunlightlabs.github.io/congress/)}
}
\seealso{
\code{\link{GoogleAPI}}, \code{\link{PopongAP}}
}
\keyword{API}

