\name{SunlightAPI}
\alias{SunlightAPI}
\title{R interface for Sunlight API}
\usage{
SunlightAPI(method, region = "korea", key = getOption("polrSunlightKey"))
}
\description{R interface for Sunlight API}
\arguments{
  \item{method}{method chooses }
  \item{region}{region is for }
  \item{key}{key is the target}
}

\value{
}
\details{
See
\url{(https://sunlightlabs.github.io/congress/) ([register](http://sunlightfoundation.com/api/)}
}
\seealso{ \code{GoogleAPI}, \code{PopongAPI}  }

\examples{

}

\keyword{API}
